feat: replace legacy toast with new notification toast
