+++ src/components/NotificationToast/index.tsx
@@ -1,0 +1,4 @@
+export type { NotificationToastProps } from './NotificationToast';
+export { NotificationToast } from './NotificationToast';
+export type { TransactionToastProps } from './TransactionToast';
+export { TransactionToast } from './TransactionToast';
