--- src/components/TransactionToast/index.tsx
@@ -1,2 +1,0 @@
-export type { TransactionToastProps } from './TransactionToast';
-export { TransactionToast } from './TransactionToast';
