+++ src/components/index.tsx
@@ -12,6 +12,8 @@ export type { LoanPositionsTableProps } from './LoanPositionsTable';
 export { LoanPositionsTable } from './LoanPositionsTable';
 export type { NotificationsPopoverProps } from './NotificationsPopover';
 export { NotificationsPopover } from './NotificationsPopover';
+export type { NotificationToastProps, TransactionToastProps } from './NotificationToast';
+export { NotificationToast, TransactionToast } from './NotificationToast';
 export type { PlusDividerProps } from './PlusDivider';
 export { PlusDivider } from './PlusDivider';
 export type { PoolsTableProps } from './PoolsTable';
@@ -22,5 +24,3 @@ export { ToastContainer } from './ToastContainer';
 export * from './TransactionDetails';
 export type { TransactionFeeDetailsProps } from './TransactionFeeDetails';
 export { TransactionFeeDetails } from './TransactionFeeDetails';
-export type { TransactionToastProps } from './TransactionToast';
-export { TransactionToast } from './TransactionToast';
