+++ src/utils/hooks/transaction/use-transaction-notifications.tsx
@@ -5,7 +5,7 @@ import { useDispatch } from 'react-redux';
 import { updateTransactionModal } from '@/common/actions/general.actions';
 import { TransactionModalData } from '@/common/types/util.types';
 import { EXTERNAL_PAGES, EXTERNAL_URL_PARAMETERS } from '@/utils/constants/links';
-import { NotificationToast, NotificationToastAction, useNotifications } from '@/utils/context/Notifications';
+import { NotificationToastAction, NotificationToastType, useNotifications } from '@/utils/context/Notifications';
 
 import { TransactionActions, TransactionStatus } from './types';
 import { TransactionResult } from './use-transaction';
@@ -60,7 +60,7 @@ const useTransactionNotifications = ({
     // creating or updating notification
     if (toastInfo.isOnScreen) {
       const toastAction: NotificationToastAction = {
-        type: NotificationToast.TRANSACTION,
+        type: NotificationToastType.TRANSACTION,
         props: {
           variant: status,
           url,
