+++ src/utils/hooks/use-sign-message.ts
@@ -1,14 +1,15 @@
 import { PressEvent } from '@react-types/shared';
 import { useCallback, useEffect } from 'react';
+import { useTranslation } from 'react-i18next';
 import { useMutation, useQuery, useQueryClient } from 'react-query';
 import { useDispatch } from 'react-redux';
-import { toast } from 'react-toastify';
 
 import { showSignTermsModalAction } from '@/common/actions/general.actions';
 import { TERMS_AND_CONDITIONS_LINK } from '@/config/relay-chains';
 import { SIGNER_API_URL } from '@/constants';
 import { KeyringPair, useSubstrateSecureState } from '@/lib/substrate';
 
+import { NotificationToastType, useNotifications } from '../context/Notifications';
 import { signMessage } from '../helpers/wallet';
 import { LocalStorageKey, TCSignaturesData, useLocalStorage } from './use-local-storage';
 
@@ -50,7 +51,9 @@ type UseSignMessageResult = {
 };
 
 const useSignMessage = (): UseSignMessageResult => {
+  const { t } = useTranslation();
   const queryClient = useQueryClient();
+  const notifications = useNotifications();
 
   const dispatch = useDispatch();
   const [signatures, setSignatures] = useLocalStorage<TCSignaturesData>(LocalStorageKey.TC_SIGNATURES);
@@ -96,17 +99,22 @@ const useSignMessage = (): UseSignMessageResult => {
     queryFn: () => selectedAccount && getSignature(selectedAccount)
   });
 
-  // TODO: add new notification
   const signMessageMutation = useMutation((account: KeyringPair) => postSignature(account), {
     onError: (_, variables) => {
       setSignature(variables.address, false);
-      toast.error('Something went wrong!');
+      notifications.show(variables.address, {
+        type: NotificationToastType.STANDARD,
+        props: { variant: 'error', title: t('notifications.signature_submission_failed') }
+      });
     },
     onSuccess: (_, variables) => {
       setSignature(variables.address, true);
       dispatch(showSignTermsModalAction(false));
       refetchSignatureData();
-      toast.success('Your signature was submitted successfully.');
+      notifications.show(variables.address, {
+        type: NotificationToastType.STANDARD,
+        props: { variant: 'success', title: t('notifications.signature_submission_successful') }
+      });
     }
   });
